Refactoring tableview cell
